Refactor HotwordDetectionServiceProximityTest
Currently, the voice cts communication across three packages. It
makes the test hard to read and we need to use broadcast, service
to let the data can pass between these thress packages.

To reduce test complexability, we make the test communication in
the same test package that can access service variables or methods
easier.

This change refactors the HotwordDetectionServiceProximityTest to
use the new infrastructure and delete the files in out service
package. There still can move common code of this test case and
HotwordDetectionServiceBasicTest to a base class, this will be
done in the follow up change.

Bug: 249287756
Bug: 263687712
Test: atest CtsVoiceInteractionTestCases
Change-Id: Ice1944c81f05cfe641353d86da3b35e884775955