Add exported flag to exposed test runtime receiver
Android T allows apps to declare a runtime receiver as not exported
by invoking registerReceiver with a new RECEIVER_NOT_EXPORTED flag;
receivers registered with this flag will only receive broadcasts from
the platform and the app itself. However to ensure developers can
properly protect their receivers, all apps targeting U
and registering a receiver for non-system broadcasts must
specify either the exported or not exported flag when invoking
registerReceiver; if one of these flags is not provided, the
platform will throw a SecurityException. This commit updates
the receiver in NetworkSecurityPolicyTestBase with the
RECEIVER_EXPORTED flag since this action is expected from the
downloads provider.

Fixes: 262909982
Test: atest NetworkSecurityPolicyCleartextUnspecifiedTest
Change-Id: I163943f646094c0a8e89a8834ebd96a0784a53f8