Allow compatibility-device-util-axt to be used with ExtendedMockito
A new target is created that does not depend on dexmaker. The new target
still needs to depend on "mockito" so that it can be built. The user of
the new target can either use "mockito-target-minus-junit4" or
"mockito-target-extended-minus-junit4" and it should work for both
dependencies.

Bug: 265378635
Test: atest SdkSandboxFrameworkUnitTests
Change-Id: Iafa7fc23681a251adc17d21057908d8f0ca6f75d
Merged-In: Iafa7fc23681a251adc17d21057908d8f0ca6f75d