Keystore: Adding CTS to check uniqueness of keys
Added CTS test cases to check uniqueness of symmetric and asymmetric
keys.
For symmetric keys same message is encrypted for 10 times and verified
that generated cipher text are unique.
For Asymmetric keys same message is signed by 10 different keys and
verified that each signature is unique.
For key agreement 10 different keys are generated and calculated a
secret for each key, and verified these secrets are uniques for 10 keys.

Bug: 221237552
Test: atest CtsKeystoreTestCases:android.keystore.cts.KeyPairGeneratorTest#testUniquenessOfEd25519Keys
Test: atest CtsKeystoreTestCases:android.keystore.cts.KeyPairGeneratorTest#testUniquenessOfRsaKeys
Test: atest CtsKeystoreTestCases:android.keystore.cts.KeyPairGeneratorTest#testUniquenessOfEcKeys
Test: atest CtsKeystoreTestCases:android.keystore.cts.KeyPairGeneratorTest#testUniquenessOfX25519Keys
Test: atest CtsKeystoreTestCases:android.keystore.cts.KeyPairGeneratorTest#testUniquenessOfEcdhKeys
Test: atest CtsKeystoreTestCases:android.keystore.cts.KeyGeneratorTest#testUniquenessOfHmacKeys
Test: atest CtsKeystoreTestCases:android.keystore.cts.KeyGeneratorTest#testUniquenessOfAesKeys
Change-Id: Ic8ababf36a5c2489151c453e6c5ac01d8e39c26f