Stop disabling bluetooth at each test
Turning off bluetooth can take up to 10 secondes, making the test suite
very slow.
There is also no need to turn the bluetooth off at each test. Instead we
want to shutdown the bluetooth at the end of all tests.

There are tests that assert the behavior of bluetooth for when it's off
and these tests will continue turning the bluetooth off. Meaning we
still need to enable the bluetooth in setup (it's a no-op if already
started)

Bug: 234892968
Test: atest CtsBluetoothTestCases
Change-Id: Ic43c3a55d4e3da7b6986421252da5514170070dd