Adding new rule for device flag + latch
The new rule ensures that when the system feature flag is true, but the
device flag is false, we ignore the tests. Over time, most devices will
eventually have a null or true device flag, but for now we require this
addition. On top of that, the latch change should fix any issues faced
by null pointer exceptions. NOTE that when flags are set to false, the
tests are (rightly) ignored. So you will notice the error rate drop to
zero, but you will not see success rate climb.

Bug: 265783728
Bug: 265553589
Bug: 265522388
Bug: 265046295
Test: These are CTS Tests.
Change-Id: I379f6823aa8b3e024c43adc684a35ca292b9f189