webkit: Remove TestWebServer from webkit libs
Two web servers were being maintained for
the webkit tests. TestWebServer is only referenced
by two tests in CTS so it makes sense to remove it.

Carried over the missing API methods those tests
were relying on that weren't present in CTSTestServer.

CTSTestServer assumes it is dealing with assets.
Introducing the setResponse APIs breaks this
assumption. As a result, the getLastRequest API method
name had to be updated to make it clear what it was doing.

Test: atest CtsWebkitTestCases
Fixes: 266048049
Change-Id: I612dc115187a84af849e0d86769d92594020204b