ITS: add circle fill metric
Details: partner device FAILed scene4 test due to the inclusion of
unfilled circles that were within absolute tolerances. Added a metric to
check that the proportion of points along a shape's horizontal and
vertical axes that match the desired color is close to 1.

Other changes: used math.isclose() instead of numpy.isclose()

Bug: 265720817
Test: tested locally on Pixel device, pending partner test.
Change-Id: I739a79f902a85b555f25ce6d0fc33447a99e3420