Fixed breakage of VENDOR_EXTENSION CTS test.
The CTS test for permission CONTROL_DISPLAY_UNITS and VENDOR_EXTENSION
was breaking because granting the VENDOR_EXTENSION permission allowed
the CarPropertyManager to expose all the vendor properties, which we are
not testing in this test. The fix was to filter out the non-vendor
properties using a bitmask and test on those only.

Bug: 262920313
Bug: 265919334
Test: atest CarPropertyManagerTest
Change-Id: Id266fce594c1ac620c8e3939f7e2faca9e9e2dd1