Wait for properties and add additional logging
This test is flakey due to a suspected race condition.  The race
condition has only been observed when running the presubmit full test
mapping.  It is not reproducible locally.

Waiting for the properties to be set may make the race condition less
likely.  Adding additional logging will make debugging this test easier
in the future.

Bug: 264937051
Test: Ran atest
Change-Id: Ic84489a2774c779cdc328a70d7c616838fb0bf29