Small fixes to UserReference
Fix UserReference#remove to check for organization owned profile owner and remove it first.
Fix UserReference#setLockCredential to correctly form the shell command with the user id before the password operand.
Fix SetPin1591StepAutomation to press the "done" button if it is present, when the user is promted to set notification preferences when setting work profile lock

Test: manual
Change-Id: Id4857a3bc80d9807aa67a46519b1165489a73b8c