Fix a few more flakiness problems in StreamedAppClipboardTest
Occasionally the input events we inject to tap on a test activity to
give it focus get rejected by InputDispatcher because it is not quite
ready for them yet, so this CL looks at the return value of injection
and retries a few times.

Also in a few cases where we're testing switching which activity is on
top and then tapping to focusing the second activity, occasionally our
injected taps would go to the first activity because it hadn't quite
been occluded by the second one yet.

Bug: 266681812
Test: atest StreamedAppClipboardTest
Change-Id: Ibeacc636f55f7f7cf28a96aca50639d1059c95f8