Clear device state before launching input activity
We have some sporadic timeouts during tests; it would appear that
a number of them are caused by the previous test not resetting the
device state. For example, dropping the notification shade will
cause this particular issue to reproduce, as the shade will take
the input events that are supposed to go to the test activity.

After investigating, other tests seem to do some device setup,
ensuring it's awake, unlocked, and the notifications are out;
I'm applying the same pattern to the hardware tests.

Bug: 266108817
Test: ran `atest SonyDualshock3UsbTest#testAllMotions` with the
notification shade down

Change-Id: I2e97b15ea31a8aff1cbfb905c96e7281a4fe9bed