Reset StrictMode setting when test is done
Currently, VelocityTrackerTest requires StrictMode to be enabled, in
order to catch any regressions in the class functionality like using the
incorrect Context.

However, before this CL, the StrictMode was getting enabled, and not
reset to the original state. That means it would also get applied to
some of the other test code.

We should look into the violations separately. For now though, only
enable strict mode for this test.

Bug: 266793271
Test: atest VelocityTrackerTest
Test: atest CtsViewTestCases
Change-Id: Iee53e71f26e41289f6c22add4c969e7bac5898ee