DeviceState - set mOriginalSwitchedUser before starting the tests.
While doing teardown and removing users that are created for a test,
DeviceState should switch to the original current user before removing
them. Otherwise if in the implementation of the test method, it is
forgotten to switch back to original current user, that test user
would be the current user and hence can not be removed, which leaves
the device in an unclean state for the next tests. This CL fixes that
issue.

Bug: 267676059
Test: atest android.multiuser.cts.UserManagerTest#setBootUser_providedUserIsNotSwitchable_nonHsum
Change-Id: I27a9188a3553a498eeb8292df8811eb49d6ee81e