speech: Speech recognition world level info public API CTS
Following entities are added as part of this change:
 - RecognitionPart data class;
 - "android.speech.extra.REQUEST_WORD_TIMING"
and "android.speech.extra.REQUEST_WORD_CONFIDENCE"
key entries for RecognizerIntent extras;
 - "recognition_parts" key entry for SpeechRecognizer result Bundle.

A recognition part represents a single word with potential adjacent
punctuation, that is returned by the SpeechRecognizer. Recognizer
result bundle contains the result in form of a list
of recognition parts under the key recognition_parts.

The details provided for a recognition part are:
 - raw text - in form of a string, always available;
 - formatted text - in form of a string, available if
"android.speech.extra.ENABLE_FORMATTING" is set in the intent;
 - timestamp - offset of the beginning of this part from the start of
the recognition session in milliseconds, available if
"android.speech.extra.REQUEST_WORD_TIMING" is set in the intent;
 - confidence level - one of the five levels of confidence, available
if "android.speech.extra.REQUEST_WORD_CONFIDENCE" is set in the intent.

Bug: 268024544
Test: A new CTS
Change-Id: Ia99917a5de02b238127f047a01acdbc35aeaab8c