Defer the removal of current user after being switched away.
While doing teardown and removing users that are created for a test,
DeviceState should call removeWhenPossible, because the user might be
the current foreground user, which will be removed when switched away.

This CL also reverts the change of orders done in ag/21224840.
Recreating the users that are removed by the tests should be done
after removing the created users, as it was before. Otherwise, we can
hit the max-users limit while creating the removed users before
removing the created ones.

Bug: 267676059
Test: atest android.multiuser.cts.UserManagerTest#setBootUser_providedUserIsNotSwitchable_nonHsum
Change-Id: I04a7ce3b8ea5b7b059df67dae7559c0e491def63