CameraITS: Create new tests for UV swap during reprocessing.
Tests whether a UV swap has occurred by comparing the output without
reprocessing with both a UV-swapped and original copy of the reprocessed
image. The metric used to compare is the Sum of Absolute Differences,
which will generally be higher for a UV-swapped image than one with the
correct U and V channels.

This test is performed separately for noise reduction, edge enhancement,
and jpeg (orientation) reprocess keys.

Bug: 266988976
Test: Ran on affected device
Change-Id: Ie013df90220c725724ae0ec2d98b050544cfb422