Deflake test.
This also modifies a few helper methods and consolidates them to be in
the same superclass.

The design of the test is inherently flaky and this fixes it for the
runs I executed locally, but there is no guarantee the test will pass
every time. But if the timing runs into those dagerous cases the
additional assertions should make the cause obvious, so that further
improvment should be easy.

Bug: 267661511
Test: atest "CtsWindowManagerBackgroundActivityTestCases:android.server.wm.BackgroundActivityLaunchTest#testActivityBlockedWhenLaunchedAfterHomePress" --iterations 100
Change-Id: I3208c6ac720c84fb2326c214259ce6f152026e19