Use multiple tasks flag to run ShareIdentityTest
The ShareIdentityTest verifies that an app only has access to the
identity of another app that launched an activity or sent it a
broadcast if the sending app opted-in to sharing its identity. To
drive the tests, ShareIdentityTest will launch a test activity in
the android.app.cts.testshareidentity package that can drive the
test per the requirements of the test case. The test activity is
launched with the new task flag which will reuse an existing
activity if one matches the Intent; this can result in a race
condition where the new activity is restarted right before
onStop / onDestroy are called against it. This commit avoids this
race condition by using the multiple tasks flag which will always
start a new activity even if an existing activity matches the
Intent.

Test: atest ShareIdentityTest --iterations 100
Fixes: 268427475
Fixes: 268072436
Fixes: 268072297
Change-Id: Ia461e9938979f2399f6e09552c452f79befdc835