Add signature verification test APKs as dependencies to appsecurity
The PkgInstallSignatureVerificationTest is intended to verify the
platform properly handles verification of all supported signature
schemes and signing algorithms. To test this, the host side test
has a dependency on a number of android_test_helper_apps; this
commit updates the appsecurity hostside test Android.bp to include
all of the test APKs as data.

Bug: 267235935
Test: atest PkgInstallSignatureVerificationTest
Change-Id: I18044ecf739c327e3d02890185d462f12f6028f9