webkit/sdksandbox: Enable pointer sync tests
We can use UiAutomation to work around the
security boundaries imposed when trying to
send pointer sync events with an instrumented process.
It didn't work for SdkSandbox tests because the SDK
process with the actual WebView is not instrumented.

This API only works from Android S onwards
but that's fine since the SDK tests only
exist from Android T.

The keyboard events don't work yet with
SurfaceView's so we'll have to keep skipping
those tests until that is fixed.

Test: atest CtsSdkSandboxWebkitTestCases
Test: atest CtsWebkitTestCases
Bug: 260196711
Change-Id: Icce9f6e7dd327d994f17b662812780ba0cde636a