Remove current user when possible and switch back to original user.
Currently if in the implementation of a test method, it is
forgotten to switch back to original current user, that test user
would be the current user and hence can not be removed during the tear
down stage, which leaves the device in an unclean state for the next
tests.

This CL sets mOriginalSwitchedUser in DeviceState before starting a
test to allow switching back to it during teardown stage. Also adds a
removeWhenPossible method to UserReference and calls it instead of
removeUser if the user to be removed is the current foreground user.
Because it's not allowed to remove the current foreground user,
but with removeWhenPossible, that user is marked ephemeral and
eventually removed after swithing to mOriginalSwitchedUser.

Bug: 267676059
Bug: 279439174
Test: atest android.multiuser.cts.UserManagerTest#setBootUser_providedUserIsNotSwitchable_nonHsum
Change-Id: Ieab24cb622d16cee16c97b89fdbc7c069c71ca82
(cherry picked from commit b676fc97193639dbbbfc06e8c8bb39b1e85fceaf)
Merged-In: Ieab24cb622d16cee16c97b89fdbc7c069c71ca82