Add getPreferredUpdateRateNanos() check
If a device does not support ADPF sessions,
getPreferredUpdateRateNanos() will return -1. We use this
to check if we should skip this test or not in `assume`.

If the assumption fails, we skip this test by returning
early.

We also need to check the test result from AtomTests.testCreateHintSession
whether it's assumption or assertion failure.

Bug: 267672503
Test: atest PerformanceHintManagerStatsTests
Change-Id: I905ec2d37a1a2d8157da78dec242c67725b937d6