Mark CtsStoragePermissionsUserDefaultSdk22 as testOnly
Update the app manifest to indicate this is a testOnly app
which allows this app to bypass the low target sdk install block.

This is necessary due to another incoming change that will remove
the --bypass-low-target-sdk-block flag when the install occurs from
a uid that is not system, root, or shell uid on a non-debug build.

Test: atest android.permission2.cts.RestrictedStoragePermissionTest
Change-Id: Ib8fe48c2353dd84ed5c0cc7513f227e9f2295796