Override window animation to 20x scale for test
And update another test to use the AutoCloseable that was created for
this since we have it now.

The 20x animation scale ensures that we pass the test on (a) devices
that have 0x animation scale set and (b) devices that are too slow
to respond to the 16ms polling check for animations to start, before
they end.

We poll for animation status in WindowState to make sure that windows
are actually animating in response to a flag saying whether they
should.

Test: atest --iterations 10 MoveAnimationTests # with rate set to 0
Test: atest --iteration 5 ActivityTransitionTests # with rate set to 0
Test: adb shell settings get global animator_duration_scale # post test
Bug: 226453118
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:30f93eca4c372c947df8e576bf4257d59acee3cd)
Merged-In: If62d06c703f882afb9157e99862db40dd9d1c224
Change-Id: If62d06c703f882afb9157e99862db40dd9d1c224