Test setProperty for write only properties
Currently we are skipping write only properties when checking if
setProperty works. This change verifies if write only properties are
able to correctly call setProperty. However, there's no way to check if
something in the vehicle actually changed because of this call, so this
is just checking that there are no unexpected exceptions thrown.

Bug: 244468655
Test: atest CtsCarTestCases:CarPropertyManagerTest
Change-Id: I57b1e23355320d8ef8d16ce0a45467dcd11a49b5