DisplayEventTest: fix flaky cached mode
When the DisplayEventTest runs in CACHED mode, it starts 3 activities as
below:

1. startActivity(DisplayEventActivity);
2. startActivity(SimpleActivity);
3. startActivity(CATEGORY_HOME);

We expect the launcher becomes the TOP, SimpleActivity becomes the LAST
(a.k.a. the previous), and DisplayEventActivity becomes CACHED.

However, it seems that Android framework can't guarantee the sequence of
step 2 and 3 above. Although step 2 starts before step 3, sometimes the
launcher in step 3 preempts the SimpleActivity in step 2. That wrongly
makes SimpleActivity the TOP and DisplayEventActivity the LAST.

Fix this by using startActivitySync in step 2 to make sure
SimpleActivity has been displayed before the launcher.

Bug: 269607576
Test: atest DisplayEventTest --rerun-until-failure
Change-Id: I0c02526a2f62884c2c0e862ecde38719390ecfea