mediav2 CTS: Rework encoder profile level test
The look up tables used by getMinLevel of different media types had
issues. This is corrected.

Updated BitStreamUtils to parse CSD/FrameHeader for profile & level
info.

The profile and level information obtained after parsing the bitstream
is compared against input configured values, values of output format and
values of extractor format of muxed file.

Some test resolutions, frame rates and bit rates are also amended basing
on the requirements mentioned in cdd 13.

As level key is only a suggestion to encoder and encoders are free to
choose a different level basing on encoding parameters, the test is
configured with min level and output of encoder is expected to contain a
level that is at least the configured value.

Bug: 151423508
Bug: 151430764
Bug: 152576008
Bug: 151398466
Bug: 166300446
Bug: 166305723
Bug: 166300448
Bug: 184889671
Bug: 184890155
Test: atest android.mediav2.cts.EncoderProfileLevelTest

Change-Id: Iad9e9b5e2a536e3f2ce7711948cbe12640112913