Fix the missing API call on phone device
This patch moved the register intent receiver to onStart to make it
paired with the unregister in the onStop, and also added flag to make
sure it can receive external app sent intents. With that change, the CTS
test FreeformWindowingModeTests can always call into the corresponding
API to test the behavior reliably.

Test: API coverage test
Test: FreeformWindowingModeTests
Bug: 264468096
Change-Id: I3649e56f8a8340e7676c2429082f58f2c3e3afb6