Enable more verbose logging for a test case to debug flaky tests
Since there are several bugs about flaky input CTS tests in
CtsHardwareTestCases that are hard to reproduce, we enable debug logs at
several stages in the input pipeline for the duration of the tests to
help debug flaky failures when they do occur. Note that this will only
work on debuggable builds.

Bug: 193231132
Test: atest GoogleAtvReferenceRemoteControlTest, and look at logcat
Change-Id: I994f7acef8ae5590b26047a6d1be0816df5cab35