Wrap rotation-dependent tests in IgnoreOrientationRequestSession
Tests that compare WindowLayoutInfo across rotation changes are
dependent on setRequestedOrientation. This worked on
a majority of devices, but large screen devices with natural
orientation being landscape and ignoreOrientationRequest=true
letterboxes activities when setRequestedOrientation(PORTRAIT).
This is a problem because WindowLayoutInfo does not return some
information for letterboxed activities (see [1]).

Tests that require rotating content and not being letterboxed
should use IgnoreOrientationRequestSession

[1] 5e0dd149001e1948cb05ff62ba4ce937a356fda6

Bug: 255240808

Test: atest android.server.wm.jetpack.ExtensionWindowLayoutComponentTest#testGetWindowLayoutInfo_configChanged_windowLayoutUpdates
Test: atest android.server.wm.jetpack.ExtensionWindowLayoutComponentTest#testGetWindowLayoutInfo_windowRecreated_windowLayoutUpdates
Test: atest android.server.wm.jetpack.ExtensionWindowLayoutComponentTest#testWindowLayoutComponent_updatesWindowLayoutFromContextAfterRotation

Change-Id: I46ca803a458a5f94f0abcee881f801d6b4244906