Set TestShareIdentityActivity's action when launching tests
The ShareIdentityTest makes use of the TestShareIdentityActivity in
a separate test package to drive verification of the activity and
broadcast identity sharing APIs; this test activity was previously
started with an explicit Intent without an action, but with the null
action Intents being blocked as part of the intent filter matching,
the Intent used to start the test could no longer find the activity.
This commit updates the test to use an explicit ACTION_MAIN to
start the test activity.

Bug: 270627143
Test: atest ShareIdentityTest
Change-Id: I5538bd8330be7f3abcfe47c19c3734fac8aa546c