Widget：Delete some test codes to avoid the failure of the operation without getting the focus
There is a CTS case testDropDownBackgroundDialogMode which tested draw
Spinner dialog menu and check showing & not showing popupwindow sending
KEYCODE_BACK. Dialog gets the window focus, but does not get the input event
processing focus, with a delay of 30ms. At this time, the keyback event is sent,
and the test activity responds to this event, resulting in failure.This test appears to be
testing the same behavior twice and we see no test coverage benefit to the flaky second
half of the test. So we think the cleanest way to deal this problem is simply to delete all the
test code after the first assertion assert(mSpinnerDialogMode.getPopupBackground()).

Test: run cts -m CtsWidgetTestCases -t android.widget.cts.SpinnerTest#testDropDownBackgroundDialogMode
Bug: 265887221
Change-Id: I349143c0f68a388b3af1fe49b8d97f560cac720d