CameraITS: Test for Display P3 JPEG.
This test captures two JPEGs: one in sRGB, and one in Display P3.
Given a chart which should produce brighter reds and greens when
capturing in P3, the ICC profiles of both images are checked and
the colors in the P3 image compared with the sRGB color space to
ensure that at least one pixel resides outside.

Tested a modified version of the test where the color space is
not explicitly set. The resulting jpeg contains an sRGB ICC
profile and passes that check. However, as no device currently
exists implementing this feature the full test can't be
validated yet.

Bug: 266754059
Test: See commit description.
Change-Id: Ib23d77f97f3ac0970657c23d632009c9343d8836