Camera CTS: Improve testCamera2AccessCallbackInSplitMode.
This test was imprecise. We only verified if we received
onCameraAccessPrioritiesChanged at least once throughout the whole focus
change sequence. A more strict expectation is necessary to avoid the
impact of false positives - onCameraAccessPrioritiesChanged could be
called for some other reason than a focus change throughout the test.

This patch enforces that the remote activity receives
onCameraAccessPrioritiesChanged and onTopResumedActivityChanged between
each tap event. This decreases the likelihood of passing on a false
positive.

Test: Ran CameraEvictionTest on UDC.
Bug: 267597879
Change-Id: I5ea6d56eaf9e5c9547d1d7de23edcf2fb3c31e8b