Remove VR branch in CTS test
We got a report of a CTS test case failing; after investigating,
it seems that the behavior of `VrManagerService` changed at some
point even though it is no longer being actively maintained.

This change will reflect the current behavior, as it's highly
unlikely that it will be fixed in the short term.

Bug: 269986910
Test: presubmit passes
Change-Id: If9a1a16b7376874984bc36401288740efcf8ab17