Add empty view to InteractiveVerifierActivity
Some CTS Verifier tests based on InteractiveVerifierActivity have no test-cases when they are opened from an automotive device.

We didn't handle this scenario gracefully and were crashing upon trying to iterate an empty list.

This fix adds an isEmpty() check and displays an empty view when there are no test-cases to display.

Fixes: 269980737
Test: run the Notification Dismissal and Notification Privacy tests from
an automotive device (or mock the isAutomotive check)

Change-Id: Id5d9f15be0899dd2f5d3fe0ffd39fe329b6ede45