Improve flaky SandboxedDetectionServiceNonExistenceTest
Currently, the code will count down the CountDownLatch first then
set the verified value. The test will verify the value when the latch
is counted down. Because the timing issue, it is possible the test
access value before the test set it. That is why we can see the flaky
result from dashboard.

In this change, we try to set value before count down CountDownLatch,
this can make sure the test can read the correct value. Also add logs
to help debugging.

Bug: 270774288
Test: atest CtsVoiceInteractionTestCases
Change-Id: I470c7481f5f318a95d18c7969c9addb0781a49c1