[GWP-ASan] [CTS] Migrate to host-side, more concrete enablement test
Migrate the GWP-ASan CTS test to be host-side. This allows future
expansion for system property usage, checking tombstones, installing
multiple apps with practically the same code (recoverable vs.
non-recoverable).

Also, make the enablement tests a little more concrete. Previously, it
was just a presence check for GWP-ASan in the mappings, now it actually
asserts that a pointer can be allocated. This "get a pointer from
GWP-ASan" will be required later for the "ensure we get a debuggerd
report" stuff.

Bug: 270720163
Test: atest CtsGwpAsanTestCases

Change-Id: I8dd724b2fd1fa84f76cf383259e764fb75de40b0