[Sharesheet CTS] Basic shortcut-target selection.
This new test simply selects a registered shortcut target and asserts
that the target activity was launched (via an intent that carries the
same shortcut ID as the registered target). Subsequent tests will
elaborate on the "alternate/refinement" logic for shortcut targets.

Note mrcasey@ has reported seeing some (pre-existing) passing CTS
tests that seem to confirm shortcut targets are visible -- even on
devices where it seems that they shouldn't be. These older tests only
*look for* the shortcut targets and stop short of *selecting* any,
but there may be bugs in the older tests, and when we investigate, we
may find that similar fixes need to be applied in the new test from
this CL.

Bug: 271149302
Test: atest CtsSharesheetDeviceTest
Change-Id: If8e8942b2faa5085dacb2cf343e402346051499b