[Sharesheet CTS] Refinement of a selected shortcut
This new test ensures that shortcut targets can be refined (by the
caller's provided refinement activity) to use any of the caller's
provided alternate intents as long as they can resolve to the same
component.

The alternates are dispatched using the same shortcut ID that was
registered for the "primary" intent. See notes in the description of
ag/21877987 concerning the (questionable) suitability of this sort of
"shortcut ID migration" to an intent other than the one that was
registered as a shortcut.

Bug: 262805893, 271149302
Test: `atest CtsSharesheetDeviceTest`. The new test fails without the
corresponding (same-topic) fix in ag/21877987.

Change-Id: Ia33b2f76fbb7598b0747db39750e508d29c81df9