Add test to verify HotwordDetectionServiceInitResultReported
Add test to verify HotwordDetectionServiceInitResultReported and
update comment wrong words.

MUST_SLEEP: need to wait CTS running finish to verify the metric
result. The metric cannot be triggered by an API call, it needs
to trigger a hotwork triger. The flow spends 15-20 secs to
complete, the test needs to wait.
Bug: 215631339
Test: atest HotwordDetectionServiceInitResultReportedStatsTest

Change-Id: Ie5af45960062c5d8d2d701e76c4e538ea6abd2c4