DO NOT MERGE: Launch SetTouchableRegionTest in fullscreen.
This test is currently failing on freeform devices due to issues in the
touchable region code (b/272786078). The bug does not occur when the
window is in fullscreen windowing mode.

Since the bug fix is nontrivial, this CL modifies the test to launch
the activity in fullscreen to unblock CTS failures. This CL also
replaces ActivityTestRule with ActivityScenario in order to specify the
launch windowing mode.

Bug: 269707702
Test: atest SetTouchableRegionTest
Change-Id: I6f97133939f5720f82eace303b8d4b6be67281c1