Add ITS test for autoframing
The idea behind the new test is to do a large zoom in a scene with faces
such that none is visible in the fov, do autoframing, wait for the state
to converge and make sure all the faces are found.

Bug: 266987960
Test: Currently since there is no HAL implementation, this test is
expected to fail. However to verify the whole code path, I set
mConvergedAutoframing (which tracks if the autoframing state has been
set to converged) to be true in
AutoframingResultListener::onCaptureCompleted(), and commented out the
subsequent num faces assertion - which made the test pass.

Change-Id: Ibb4faf559c40837cf2af94e2bf9e0082844d1d15