Increase timeout to avoid test failures
In some of the tests in StreamedAppClipboardTests, we need to create a
new VirtualDevice and VirtualDisplay, and then inject taps on the
display to give it focus. But for a brief after creation, input
injection seems to get rejected. These tests already retry the
injection a few times up to a timeout value when this happens, but
occasionally the timeout is exceeded, particularly on cuttlefish
emulators. I haven't yet found any sort of signal we could listen for
to know when input injection is ready to be received, so I'm just
increasing the timeout value a little in this CL to see if that helps
avoid failures.

Bug: 271456818
Test: atest StreamedAppClipboardTest
Change-Id: I315c588ebe2494e68ff0353158f2c25f8d5c84f5