CameraITS:Test to verify led snapshot works ok.
This change includes adding a new test test_led_snapshot to verify
that the LED snapshot works ok. This new test is added in
scene/sensor_fusion since for the test setup tablet cannnot be used as
it is reflective. A checkerboard chart will be used for this test.
The test will skip if the camera device flash does not have flash and if the
device is below Android level 14.
Images will be captured in normal lighting settings without flash, with
dark lighting conditions auto-flash off and dark settings with
auto-flash ON.
Image mean will be computed and will be checked to be in the level.
R/G and B/G ratios will be calculated to be in the level set in the
test.

Bug: 110214159

Test: Tested on a pixel device.
Change-Id: I1cc6a95b056cdb96a9b0d6983eaa63057b269444