Deflake KeyguardTests race conditions
Reparative and preventative fix for race conditions seen in these tests
that had been caused by using computeState as a blocker for visibility
calculations. computeState only guarantees that the passed-in activities
have valid state, not that the valid state is the one we want in the end
for the assertion.

The waitAndAssertResumed / waitAndAssertStopped helpers wrap in a wait
for the right state and then assert the same things about visibility
that we need here.

Since the assertion messages that tests pass in for waitAndAssert... are
generally rather mechanical, this patch also adds a helper without the
required "expected activity X to be visible" assertion message.

Test: atest --iterations 20 android.server.wm.KeyguardTests
Bug: 272531865
Change-Id: Ia91cf38eb2ccf93b1a9b03f17b98ba01af315243