Fix NotificationPermissionTest for wear
Multiple NotificationPermissionTests fail on slower devices. The test
app needs to wait for the window to be focused before creating the
notification channel so the permission prompt can be displayed.
The test application also keeps asking for permissions because onStart
gets called when the requestPermission activity stops. Moving the code
in onStart to onCreate fixes this.

Bug: 271487257
Test: atest NotificationPermissionTest
Change-Id: I07de1d40b43f6ac74e758e6fcd19c6cd70d88ca3