Render SV at half the display size for TPL test
TrustedPresentationCallbackTests require the entire content to be
visible to pass. Hardcoding a size could fail on different devices if
the display itself is smaller and cuts off some of the content. Instead,
render the content at half the display size to ensure the content is not
cut off by the display. Also make sure the content is in the center to
ensure system bars don't cut off some of the content.

Test: TrustedPresentationCallbackTest
Fixes: 271448673
Change-Id: Ia5e63d5a196b7b24f93b3048830ac54af950b228