Separate out a test for KCAs on activity finish
If we finish an activity, the activity underneath starts showing, so the keep-clear areas from it should be expected by the test. For this,
we add a new test and save the keep-clear areas existing on the screen before we start the test activity. Once we finish the test activity, we assert that the keep-clear areas match the ones that existed previously

Bug: 272761032
Test: atest KeepClearRectsTest
Change-Id: I8a74c852c41fd421a9a7a054915be12a0b7ecabc