Add VCN CTS to verify network selection based on capabilities
This commit extends the existing network selection tests to also
cover selection based on capabilities. It also verifies that
VCN can be built on a non-Internet network.

Additionally, this commit removes capabilities that are not allowed
on a test network from the TestNetworkWrapper since those capabilities
will be removed by the ConnectivityService anyway.

Bug: 263415068
Test: atest VcnManagerTest (new tests)
Change-Id: Ibba855647d3b391ac94670b3672076c1fbb3fc61