Use task display area's windowing mode in AssistantStackTests.
testAssistantStackLaunchNewTask and
testAssistantStackFinishToPreviousApp in AssistantStackTests are failing
on freeform devices due to the assumption that the test activity's
windowing mode is the same as the default display's windowing mode.

On freeform devices, the display's windowing mode is freeform in S. This
was changed in T such that the display's windowing mode is fullscreen,
and the task display area remains as freeform. Therefore, we need to
update the tests' expectation to fit this behavior.

Bug: 269964891
Bug: 269964763
Test: atest AssistantStackTests
Change-Id: Ibf01e1af10631aa3a3a71a7f7a4f6fb6ab3c2558