Only assert that the cur profile directory is accessible.
Whether ART creates a cur profile or not is an implementation detail. It
cannot be checked by the CTS because the behavior can change, given that
ART is an updatable module. In fact, the behavior will change once ART
Service is enabled (which is a new feature for U+ only).

Bug: 274513882
Test: atest CtsAppSecurityHostTestCases:AppDataIsolationTests#testSharedAppAbleToAccessOtherAppDataDir
Change-Id: I89304508f2adb9abed158327297875f806516f25