Create VirtualDisplayActivityScenarioRule as a test helper
Add a test rule that launches a virtual display and launches the input
CTS activity on that display.

After this refactoring, it will be easier to share this test code with
different tests.

Also refactor UinputTouchScreen to be more flexible, allowing it to be
used for other touch devices like touchpads, drawing tablets, and
styluses.

Bug: 236798672
Test: atest TouchScreenTest
Change-Id: Ie23dbbe6d56f37e0f6f4d644092deab2a7a2e4a8