Update window-embedded overlay test to use a view hierarchy.
Real services using window-embedded overlays are likely to want to
render multiple overlay views on top of content-app views, such as
adding text labels like Voice Access, so this test now mimics that
scenario.

One FrameLayout is added as a window-embedded overlay that matches the
size of the content-app activity's window. Then, a subview is added
to the overlay that matches the bounds of a node within the content app.
Finally, we assert that the overlay's subview and the content-app node
have the same window-space and screen-space bounds.

Fix: 261248638
Test: atest AccessibilityOverlayTest
Change-Id: I14d6cc5b25dde91ed6407c0ce66beb3dfc397b3d