Reduce to wait service ready time from 10 seconds to 5 seconds
When running the test case, we will set the testing service.
After setting the new testing service, the original service
will be shutdown. In order to avoid the flaky case when running
the test case during the original service is shutdown. We wait
10 seconds to let the original service finish the shutdown.

After local testing 100 runs, it works well to reduce the time
from 10 seconds to 5 seconds.

Bug: 275108663
Test: atest HotwordDetectionServiceBasicTest --iterations 100
Change-Id: If48867d9e8f0d171c54b29d69cb29d1eeb0b1b9a