CTS test for optout from ViewAPI sandboxing
When OVERRIDE_SANDBOX_VIEW_BOUNDS_APIS is enabled by OEM
for an application, there is a way for the application to
opt-out from this treatment by setting the followinng property
to false in AndroidManifest.xml.

<application>
  <property android:name=
  "android.window.PROPERTY_COMPAT_ALLOW_SANDBOXING_VIEW_BOUNDS_APIS"
            android:value="false"/>
</application>

This CL adds a test that checks that this treatment gets
disabled for this application task when this property is set in
AndroidManifest.xml.

Bug: 275049701
Test: atest CtsWindowManagerDeviceTestCases:CompatChangeTests
Change-Id: I6b4d13287423acb3914ecdefaed52182426932f7