Use same contacts data setup for all test cases
This change moves the logic to insert and delete test contacts data to
the setup and teardown methods of the test class. These will only be
called once per test. The tests should not change any contacts related
data anyway since the contacts write are disabled from cloned apps. The
change also adds assertions to check if the deletion of contacts
happens successfully in the teardown.

Test: atest ContactsSharingTest
Bug: 275304476
Change-Id: I915199142b67ed0aae6ce2313a7c36813e457a57