Fix PointerIconTest
The input team recently committed a CL that prepares the pointer
cursor logic to be used also by the stylus; this broke a CTS test.

After checking the code, turns out that the CTS test sends the
source as being the touchscreen, which will never happen in real
devices; currently only the mouse events will result in pointer
icons. This CL changes the source of that `MotionEvent` to mouse.

Test: Presubmit, ABTD
Bug: 276354228
Change-Id: I01571f2c0eefc8a694d4e9dd576511a746be5d45