Bunch of headless test fixes.
Also convert "default" dpc queries so it can be used for non-dpc
defaults. Previously we hardcoded the classname inside the annotation -
but that won't work when a DPC can actually be a device admin or a
delegate etc. This gets rid of that hardcoding so a follow up CL can
allow us to specify additional query arguments when using e.g.
@PolicyAppliesTest.

Test: btest a.d.c.BugReportTest
Fixes: 276314657
Change-Id: I9c6c77430a3d731eccb56a2ce908e3c3dfe0784f