Improve NoProcessOutgoingCallPermissionTest
The test now uses a test app, that has the permission the test lacks.
The test app will send a message to the test when it receives the
outgoing call broadcast, and the test will wait a bit longer than that,
before deciding the test is complete. The test also ends the call, now.

Fixes: 276532886
Test: atest NoProcessOutgoingCallPermissionTest
Change-Id: If7c60cd344fdb7e23ca1747077c03e6d0f79b236