Attempt fixing flakiness by prevent display rotates
The test activity was started in landscape mode because the device
was positioned in landscape mode. However, the original task width
and height was retrieved from the configuration while in portrait
mode since the Launcher app was portrait only.

Adding a rotation session to prevent display rotates during the test.

Bug: 275415374
Test: ActivityEmbeddingPlaceholderTests
Change-Id: If288126ce749cd0194b8f7f270d2a92e0f8ba39b