SystemUtil: Make permission adoption composable
SystemUtil's permission adoption affects the state of the currently
adopted permissions, because it always drops all permissions in the end,
instead of restoring the permissions that were previously adopted. This
changes the permission adoption mechanism to be composable, where
calling runWithShellPermissionIdentity() will not affect the state of
adopted permissions after the call completes.

Bug: 193231132
Test: presubmit
Change-Id: Ic1b1de9149e2d87240205b90054e1981b80f59b9