Removing unnecessary code
If one delegate is destroyed in multiple delegates,  after a few seconds it internally triggers the destruction of the remaining delegates and then the recreation of another delegate with the new feature set that it supports.

We have removed the test that checks for session decrease at that moment since it was unnecessary and could cause timing errors.

Bug: 276869341
Test: cts SipDelegateManagerTest

Change-Id: I0de4efbfa971b740ec6fa0fefdecce6f71a46f41