Ensure test app isnt installed before running tests
Found flakiness in some test invocations where test app is found to be
installed in a restricted profile. It is suspected that the test app
installed from the previous test invocation, a different test, or a test
that installs an app in an unrestricted profile is lingering, causing
other tests to fail.

Bug: 277987631
Test: atest CtsPackageInstallTestCases:UserRestrictionInstallTest
Change-Id: Ie9755f3ec71748b4f458025dd6d4137450cb10af