Fix flaky testLogHotwordDetectorEventsConnectedAppUpdateState
The HotwordDetectorEvents#ON_DISCONNECTED will not be logged if the
HDS flag mRespectServiceConnectionStatusChanged is not set to false.
We found the faled case, the restart CTS will be run before the test.
It is possible causing the flag be set that makes ON_DISCONNECTED be
logged. We will have an extra event here.

Use isAtLeast that only focuses on the events we expected.

Bug: 278313422
Test: atest HotwordDetectorEventsStatsTest
Change-Id: Ibcbc11437332abac7a494c6d2e227eab2c670cc3