Add tearDown for tests that extend BaseTelecomTestWithMockServices
Tests that call super.setUp() but not super.tearDown() have registered
listeners and not removed them. This causes the CTS process to reach the
maximum number of listeners and subsequent tests will fail/crash,
especially if tests are run multiple times.
Ensure that tearDown is called for all tests that have called setUp so
that cleanup is done properly for all tests.

Test: atest CtsTelecomTestCases
Bug: 277288708
Change-Id: I52f33e105e0f92f14f1e8dd5a693bce6823c650b