Add test to ensure IMM#mCurRootView is cleared
This adds a test to ensure the mCurRootView property from
InputMethodManager gets cleared when it becomes inactive or loses window
focus.

Before, when going from an app that had the IME shown straight
to the launcher home screen, and tapping on the Google search bar,
this would attempt to show the IME but would sometimes fail. This
is because the show flow would progress further than normal, due to
the leftover mCurRootView value in the launcher process.

Clearing this value ensures the erronous flow is cancelled early, and
due to a subsequent retry, can always succeed in showing the IME.

Test: atest android.view.inputmethod.cts.FocusHandlingTest#testClearCurRootViewWhenDifferentProcessBecomesActive
Bug: 276742733
Change-Id: Id80eae19f479103c614fa04a0bda4cd6aeae5bd4