Opt out of foldable configuration runs
Currently, the test code waits the Activity finish to set the signal
to test to let it know it is started correctly. The tests fails only
when the device in the open mode, the Activity will not be finished.
Because the voice setting Activity is really started, our function
works fine. This should have some problem the Activity cannot be
finished correctly in the open mode. Because our function works and
voice settings doesn't have the different behavior for foldable device.
It is ok to opt out from foldable configuration.

Bug: 278822796
Test: atest android.voicesettings.cts
Change-Id: I6785c5108874a2053231bd9736f05fb47326b9d5