[DO NOT MERGE] Add status bar vis check in HotwordDetectionServiceBasicTest.
HotwordDetectionServiceBasicTest#testHotwordDetectionService_onDetectFromExternalSource_success
checks for the microphone indicator UI on the status bar. Therefore,
this test will fail on devices where there is no status bar window. This
CL adds a visibility check for the status bar before making the
assertion.

This is only needed to bypass failures in T; in U onwards, the test has
been modified to not rely on the UI indicator.

Bug: 278915819
Test: atest HotwordDetectionServiceBasicTest
Change-Id: I3c34332c77a2a88b79eaba274490c0f4ce82994a