Make A11yMagnificationTest magnification bounds aware
Currently many A11yMagnificationTest use a center point of the top
window bounds. However, it's not guaranteed that there's a touchable
region at that point. When there isn't, request is adjusted and the test
fails.
With this change, requested magnification config will be inside the
magnification region.

Note that this change is not needed for tests for window magnification.
But CL modifies all usages of point computations for consistency.

Also note that MagnificationController#getMagnificationRegion() is a
deprecated API. #getCurrentMagnificationRegion is a new altenative, but
it returns bounds for window magnification when the last enabled feature
is window magnification. It's more robust to use the former so that we
don't depend on the device state.

Bug: 278629301
Test: AccessibilityMagnificationTest
Change-Id: I7c041675748ff1bc1ae846a9bdf0a4cb97441832