Fix PhotoPickerBannersTest flakes and failures
The tests were flaky due to the picker activity being relaunched without
sparing enough time to create the banner controller.

Also, the test activity doesn't have access to the path on the device
where we store the cloud provider info. Due to this, the banners weren't
WAI as we were using multiple picker sessions to reset the banner data.

Test: atest PhotoPickerBannersTest
Bug: 274840171
Change-Id: I027c45a77dcf21766c65b81aec0b9891899bb613