Specify fullscreen windowing mode in FlagSlipperyTest.
The SlipperyEnterBottomActivity is expected to be launched in fullscreen
windowing mode in FlagSlipperyTest, but this is not specified when the
activity is launched. This leads to the activity being launched in
freeform windowing mode on certain devices, leading to the tests
failing. This CL adds the windowing mode specification to the activity.

Bug: 273832065
Test: atest FlagSlipperyTest
Change-Id: I017c81f3553d97184a6a9fb4827c718cee79e940