Pass CLEAR_TASK intent flag to MANAGE_APP_PERMISSION
When starting the MANAGE_APP_PERMISSION screen, pass the
FLAG_ACTIVITY_CLEAR_TASK intent flag. This will restart the screen
even if it was already started.

Before this, our CTS tests would fail if this screen was already open
(for example from a previous test) if the activity had previously
been started with different EXTRA arguments.

Test: atest CtsPermission3TestCases
Bug: 275459463
Change-Id: I5d6010f6345d8e6ed459866b5e91c332430ee64d