[Hotword] Check the uid in OnOpNotedListener
- Check the uid in OnOpNotedListener.
To avoid noise impacting the test result, check the uid is as we
expected. The test adopts shell permission to pass the permission
check, so the uid should be the uid of Shell.
- Remove the check of the op in the listener
We only start to watch on the specific op, we don't need to check
the op in the listener again.

Bug: 279574756
Test: atest HotwordDetectionServiceBasicTest
Test: atest HotwordDetectionServiceStressTest
Change-Id: Ifaf010a55c596a8d9b17afe70a80960dd6c34ce0