LocationProviderInterceptDialogTest - Don't run on auto/tv/watch
This test is currently failing on tv. But, the LocationProviderInterceptDialog is only invoked from handheld, so don't run it on auto, tv, or watch.

Also, add FLAG_ACTIVITY_CLEAR_TASK when opening the ACTION_MANAGE_APP_PERMISSIONS screen. This clears the screen if it is already open, prevent rare intermittent failures.

Test: atest LocationProviderInterceptDialogTest
Fix: 279183829
Change-Id: Ibc1335938ed26415088426560ca5e914face0088