Update os.cts.DebugTest#testPrintLoadedClasses to print less to logcat
PrintLoadedClasses tests VMDebug.printLoadedClasses that prints the
information about all loaded classes into the logcat. It seems that this
is causing problems with test infra so modifying the test to skip
printing detailed information about classes. In future, we may want to
update this test so we actually verify the output too.

Bug: 278291584
Test: atest CtsOsTestCases:android.os.cts.DebugTest#testPrintLoadedClasses
Change-Id: I4da22b183e46f1b1d38af756c3b4d3c2b17ebeb0