fix(non linear font scaling): increase test tolerance for rounding errors
This test was failing on some devices when textSize is loaded from an
attribute, because in that case the size is rounded to the nearest
pixel, which throws off these precise calculations in tests.

A few pixels doesn't make a noticeable difference to the user, but we
need a wide tolerance in tests.

Fix: 279456702
Test: atest cts/tests/tests/widget/src/android/widget/cts/TextViewFontScalingTest.kt
Change-Id: If811b346db43f45539701e03b13d70c689ecda7d