fix testWindowProducerScaling CTS tests
This tests checks if the window scales the pixels correctly to a smaller sized bitmap. Also, the Activity launched by the tests sets the flag SYSTEM_UI_FLAG_HIDE_NAVIGATION.
Now the activity window still draws behind the status bar. In Automotive OEMS can put the status bar on left/right side, width of which could be unknown.
Under the following scenario:
Activity window is 1000 x 800  (W X H)
Status bar is on the right side with width 500 (This window is on top of activity window)
Now the activity content gets inseted at decorView and ONLY gets 500 (1000 - 500) x 800 . Which means the other half of the window is black / transparent.

Now when the tests scale the window down it checks for the pixel from the content view that activity added. But as the second/right half of the window was black the bitmap replicates the same and hence tests fails.
In case of testing the window scaling the tests should launch the activity in full screen mode so that the pixels of the content covers the entire window and while scaling down it can match.

fix: 280296973
Test: atest android.view.cts.PixelCopyTest#testWindowProducerScaling
Change-Id: I981a26676c549ffc4bc3c80924d8e30ec4d12637