Fix LocationAccessCheckTest on release build.
Non-debuggable apps are disallowed to be downgraded by shell on
non-debuggable builds now, which broke this test on release builds.

We actually don't test anything related to version codes, and it's
perfectly fine to update apps with the same version code. So we should
just remove the troublesome version code in test helper app manifest.

Bug: 280741169
Test: LocationAccessCheckTest
Test: PermissionControllerTest
Change-Id: I8f7c67a1565cffa6e8545676e928ffdd87e93052