Ensure tearDown called when setUp errors on assert
Currently, when setUp errors out on an assertion, the test fails but
tearDown is not called after for cleanup. This causes unused listeners
to accumulate on the telecom cts pid, making other tests fail as well.

This error can be reproduced locally by forcing an assertion error in
setUp (i.e. CallDetailsTest) and we will see that subsequent test runs
are affected when tearDown isn't called. Catching the error and forcing
tearDown should resolve this issue so that subsequent tests aren't
affected.

This CL only addresses vulnerabilities found around asserting code logic
within setUp and tearDown in order to preserve the initial context for
why this bug was raised. Further refactor can be done for the rest of
the CTS test suite to ensure tearDown is called after setUp throws an
exception (in a separate CL). This can be combined with efforts to
optimize the existing tests (i.e. disregaring setUp logic when not
testing for telecom, consolidating common setUp logic so that it's not
run every time for each test).

Fixes: 280380186
Test: atest CtsTelecomTestCases
Change-Id: I7e901dfe7e8db337f9d91e3366db19179297b4f2