Press HOME before ToolbarActionBarTest
Bug: 261605837

This appears to fix the problems with ToolbarActionBarTest,
so it is likely that a previous test left the device
in a bad state. It may be that a notification shade was
left open or some other window or bubble notification
was left on top.

Test: ABTD for CtsAppTestCases
Change-Id: Icf781e6f9b05b39be33393bbcfc1af368dee0eee
(cherry picked from commit 50143ba8f06a63f9d4b79411072dcfa3acc44962)