Enhance existing call/connection event tests with Bundles
When testing call/connection events in CTS, we never sent
information in the test Bundles themselves, which means that if
they were somehow altered in Telecom when they werent supposed
to be, CTS testing never caught the issue.

This change ensures that relatively complex bundles including
custom Parcelables and IBinder references are successfully delivered
when sent as an event through Telecom. This ensures that extension
APIs that are built on the event passing mechanism provided by call
and connection events are properly supported in the platform.

Fixes: 281588944
Test: atest CtsTelecomTestCases:TransactionalApisTest and CallDetailsTest
Change-Id: I92a0b01d5c5bf16521e75a1f6f1d0d7a32241dc1